eiei

+++ controller/editSubject.php
@@ -12,18 +12,28 @@
     session_start();
     if(isset($_SESSION['user'])){
         $userlogin = $_SESSION['user'];
-    }
-    $dataSubjects = array();
-    $dataSubjects = getListSubject();
-    if(isset($_POST["edit"])){
-        $id = $_POST["edit"];
+        if($userlogin->getTypeUser()=="ADMIN") {
+            $dataSubjects = array();
+            $dataSubjects = getListSubject();
+            if (isset($_POST["edit"])) {
+                $id = $_POST["edit"];
 
+            }
+            if (isset($_GET["idSubject"])) {
+                $id = $_GET["idSubject"];
+                $de = deleteSubject($id);
+            }
+            include_once("../view/edit_subject.php");
+        }
+        else{
+            header("Location: ../index.php");
+            exit();
+        }
     }
-    if(isset($_GET["idSubject"])) {
-        $id = $_GET["idSubject"];
-        $de = deleteSubject($id);
+    else{
+        header("Location: ../index.php");
+        exit();
     }
-    include_once ("../view/edit_subject.php");
 ?>
